Avoid exporting range check
