+++ sqlite3/lib/src/common/functions.dart
@@ -144,12 +144,3 @@ class AllowedArgumentCount {
   const AllowedArgumentCount(this.allowedArgs);
   const AllowedArgumentCount.any() : allowedArgs = -1;
 }
-
-extension BigIntRangeCheck on BigInt {
-  BigInt get checkRange {
-    if (this < bigIntMinValue64 || this > bigIntMaxValue64) {
-      throw Exception('BigInt value exceeds the range of 64 bits');
-    }
-    return this;
-  }
-}
