+++ sqlite3/lib/src/common/impl/utils.dart
@@ -42,3 +42,12 @@ int eTextRep(bool deterministic, bool directOnly) {
 
   return flags;
 }
+
+extension BigIntRangeCheck on BigInt {
+  BigInt get checkRange {
+    if (this < bigIntMinValue64 || this > bigIntMaxValue64) {
+      throw Exception('BigInt value exceeds the range of 64 bits');
+    }
+    return this;
+  }
+}
