+++ sqlite3/lib/src/ffi/impl/implementation.dart
@@ -4,7 +4,7 @@ import 'dart:typed_data';
 
 import '../../../sqlite3.dart' hide sqlite3;
 import '../../common/constants.dart';
-import '../../common/impl/database.dart';
+import '../../common/impl/utils.dart';
 import '../ffi.dart';
 
 part 'data_change_notifications.dart';
